+++ tests/LatexTest.java
@@ -16,7 +16,7 @@ public class LatexTest {
 
 	@Test
 	void testVar() {
-		GeneralFunction test = FunctionParser.parseInfix("\\pi\\Gamma");
+		GeneralFunction test = FunctionParser.parseInfix("\\pi \\Gamma");
 		assertEquals(10, test.evaluate(Map.of("Γ", 3.2)), .3);
 	}
 
@@ -60,5 +60,17 @@ public class LatexTest {
 	void spaceEscapedLatex() {
 		GeneralFunction test = FunctionParser.parseInfix("1 + \\sin(x)");
 		assertEquals(1, test.evaluate(Map.of("x", 3.0)), .3);
+	}
+
+	@Test
+	void adjacentWords1() {
+		GeneralFunction test = FunctionParser.parseInfix("\\pi\\Gamma");
+		assertEquals(6.3, test.evaluate(Map.of("Γ", 2.)), .3);
+	}
+
+	@Test
+	void adjacentWords2() {
+		GeneralFunction test = FunctionParser.parseInfix("\\epsilon\\Gamma\\epsilon\\pi");
+		assertEquals(6.3, test.evaluate(Map.of("ϵ", 1., "Γ", 2.)), .3);
 	}
 }